# artifacts/api-server/src/lib/auth.ts
import {
  createHmac,
  randomBytes,
  scrypt as scryptCallback,
  timingSafeEqual,
} from "node:crypto";
import { promisify } from "node:util";
import type { NextFunction, Request, Response } from "express";
import { and, eq, gt } from "drizzle-orm";
import { db, sessionsTable, usersTable } from "@workspace/db";
import { logger } from "./logger";

const scrypt = promisify(scryptCallback);
const SESSION_COOKIE = "reair_session";
const SESSION_DAYS = 30;

export const USER_ROLES = ["admin", "editor", "viewer"] as const;
export type UserRole = (typeof USER_ROLES)[number];

export type CurrentUser = {
  id: number;
  email: string;
  role: UserRole;
  createdAt: Date;
};

export function normalizeUserRole(role: string): UserRole {
  return USER_ROLES.includes(role as UserRole) ? role as UserRole : "viewer";
}

export function effectiveUserRole(user: Pick<CurrentUser, "email" | "role">): UserRole {
  const configuredEmail = process.env.ADMIN_EMAIL?.trim().toLowerCase();
  return configuredEmail && user.email === configuredEmail
    ? "admin"
    : normalizeUserRole(user.role);
}

export function isAdministrator(user: Pick<CurrentUser, "email" | "role">): boolean {
  return effectiveUserRole(user) === "admin";
}

export function canEditReports(user: Pick<CurrentUser, "email" | "role">): boolean {
  return effectiveUserRole(user) !== "viewer";
}

function hashToken(token: string): string {
  const secret =
    process.env.SESSION_SECRET ?? "development-only-session-secret";
  return createHmac("sha256", secret).update(token).digest("hex");
}

export async function hashPassword(password: string): Promise<string> {
  const salt = randomBytes(16);
  const derived = (await scrypt(password, salt, 64)) as Buffer;
  return `scrypt$${salt.toString("hex")}$${derived.toString("hex")}`;
}

export async function verifyPassword(
  password: string,
  encoded: string,
): Promise<boolean> {
  const [, saltHex, hashHex] = encoded.split("$");
  if (!saltHex || !hashHex) return false;

  const derived = (await scrypt(password, Buffer.from(saltHex, "hex"), 64)) as Buffer;
  const expected = Buffer.from(hashHex, "hex");
  return derived.length === expected.length && timingSafeEqual(derived, expected);
}

export async function seedAdminUser(): Promise<void> {
  const configuredEmail = process.env.ADMIN_EMAIL?.trim().toLowerCase();
  const configuredPassword = process.env.ADMIN_PASSWORD;

  if (!configuredEmail && !configuredPassword) return;
  if (!configuredEmail) {
    throw new Error("ADMIN_EMAIL is required when ADMIN_PASSWORD is set.");
  }
  if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(configuredEmail)) {
    throw new Error("ADMIN_EMAIL must be a valid email address.");
  }

  const existing = await db
    .select({ id: usersTable.id, role: usersTable.role })
    .from(usersTable)
    .where(eq(usersTable.email, configuredEmail))
    .limit(1);

  if (existing[0]) {
    if (existing[0].role !== "admin") {
      await db
        .update(usersTable)
        .set({ role: "admin" })
        .where(eq(usersTable.id, existing[0].id));
    }
    logger.info({ email: configuredEmail }, "Configured admin account already exists");
    return;
  }
  if (!configuredPassword || configuredPassword.length < 8) {
    throw new Error("ADMIN_PASSWORD must be at least 8 characters long when seeding a new administrator.");
  }

  await db
    .insert(usersTable)
    .values({
      email: configuredEmail,
      passwordHash: await hashPassword(configuredPassword),
      role: "admin",
    })
    .onConflictDoNothing({ target: usersTable.email });

  logger.info({ email: configuredEmail }, "Seeded configured admin account");
}

export async function createSession(
  userId: number,
  response: Response,
): Promise<void> {
  const token = randomBytes(32).toString("hex");
  const expiresAt = new Date(Date.now() + SESSION_DAYS * 24 * 60 * 60 * 1000);

  await db.insert(sessionsTable).values({
    tokenHash: hashToken(token),
    userId,
    expiresAt,
  });

  response.cookie(SESSION_COOKIE, token, {
    httpOnly: true,
    sameSite: "lax",
    secure: process.env.NODE_ENV === "production",
    maxAge: SESSION_DAYS * 24 * 60 * 60 * 1000,
    path: "/",
  });
}

export async function getCurrentUser(request: Request): Promise<CurrentUser | null> {
  const token = request.cookies?.[SESSION_COOKIE] as string | undefined;
  if (!token) return null;

  const rows = await db
    .select({
      id: usersTable.id,
      email: usersTable.email,
      role: usersTable.role,
      createdAt: usersTable.createdAt,
      sessionId: sessionsTable.id,
    })
    .from(sessionsTable)
    .innerJoin(usersTable, eq(usersTable.id, sessionsTable.userId))
    .where(
      and(
        eq(sessionsTable.tokenHash, hashToken(token)),
        gt(sessionsTable.expiresAt, new Date()),
      ),
    )
    .limit(1);

  if (!rows[0]) return null;
  return {
    id: rows[0].id,
    email: rows[0].email,
    role: effectiveUserRole({
      email: rows[0].email,
      role: normalizeUserRole(rows[0].role),
    }),
    createdAt: rows[0].createdAt,
  };
}

export async function destroySession(request: Request, response: Response): Promise<void> {
  const token = request.cookies?.[SESSION_COOKIE] as string | undefined;
  if (token) {
    await db.delete(sessionsTable).where(eq(sessionsTable.tokenHash, hashToken(token)));
  }
  response.clearCookie(SESSION_COOKIE, { httpOnly: true, sameSite: "lax", path: "/" });
}

export async function requireUser(
  request: Request,
  response: Response,
  next: NextFunction,
): Promise<void> {
  const user = await getCurrentUser(request);
  if (!user) {
    response.status(401).json({ error: "Authentication required" });
    return;
  }
  request.currentUser = user;
  next();
}

export function requireAdministrator(
  request: Request,
  response: Response,
  next: NextFunction,
): void {
  if (!request.currentUser || !isAdministrator(request.currentUser)) {
    response.status(403).json({ error: "Administrator access required" });
    return;
  }
  next();
}

export function requireReportEditor(
  request: Request,
  response: Response,
  next: NextFunction,
): void {
  if (!request.currentUser || !canEditReports(request.currentUser)) {
    response.status(403).json({ error: "Editor or administrator access required" });
    return;
  }
  next();
}

declare global {
  namespace Express {
    interface Request {
      currentUser?: CurrentUser;
    }
  }
}