# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface User {
  id: number;
  email: string;
  createdAt: string;
  isAdmin: boolean;
  role: UserRole;
}

export interface AuthCredentials {
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
}

export type CreateUserRequest = AuthCredentials & {
  role: UserRole;
};

export interface UpdateUserRoleRequest {
  role: UserRole;
}

export interface AuthSession {
  user: User;
}

/**
 * @nullable
 */
export type SessionStateUser = {
  id: number;
  email: string;
  createdAt: string;
  isAdmin: boolean;
  role: UserRole;
} | null;

export interface SessionState {
  authenticated: boolean;
  /** @nullable */
  user: SessionStateUser;
}

export interface ReportUpload {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  content: string;
}

export interface Report {
  id: number;
  name: string;
  clipCount: number;
  uploadedAt: string;
}

export interface Note {
  tc: string;
  /** @nullable */
  secs: number | null;
  text: string;
}

export interface Clip {
  id: string;
  reportId: number;
  source: string;
  /** @nullable */
  date: string | null;
  /** @nullable */
  revision: string | null;
  /** @nullable */
  time: string | null;
  /** @nullable */
  originalAir: string | null;
  /** @nullable */
  lastAir: string | null;
  hosts: string[];
  guests: string[];
  shortSynopsis: string;
  longSynopsis: string;
  duplicateLongSynopsis: boolean;
  sensitiveNotes: Note[];
  dateNotes: Note[];
  flagCount: number;
}

export interface DashboardSummary {
  reportCount: number;
  clipCount: number;
  reviewCount: number;
  flagCount: number;
  /** @nullable */
  earliestClipDate: string | null;
  /** @nullable */
  latestClipDate: string | null;
}