# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Account access
  - name: reports
    description: Uploaded report management
  - name: clips
    description: Parsed clip reports
  - name: dashboard
    description: Workspace summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current signed-in user
      responses:
        "200":
          description: Current session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionState"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Signed-in session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid email or password
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out
      responses:
        "204":
          description: Signed out
  /auth/users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List users for the administrator
      security:
        - cookieAuth: []
      responses:
        "200":
          description: User accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Administrator access required
        "401":
          description: Authentication required
    post:
      operationId: createUser
      tags: [auth]
      summary: Create a user account for the administrator
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid or duplicate user
        "403":
          description: Administrator access required
        "401":
          description: Authentication required
  /auth/users/{userId}:
    patch:
      operationId: updateUserRole
      tags: [auth]
      summary: Update a user's role
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid role or protected administrator
        "403":
          description: Administrator access required
        "401":
          description: Authentication required
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      tags: [auth]
      summary: Delete a user account for the administrator
      security:
        - cookieAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: User deleted
        "400":
          description: The current administrator cannot delete their own account
        "403":
          description: Administrator access required
        "401":
          description: Authentication required
        "404":
          description: User not found
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List uploaded reports
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Reports in the shared workspace archive
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "401":
          description: Authentication required
    post:
      operationId: uploadReport
      tags: [reports]
      summary: Upload and parse a CSV report
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpload"
      responses:
        "201":
          description: Uploaded report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Report could not be parsed
        "403":
          description: Editor or administrator access required
        "401":
          description: Authentication required
  /reports/{reportId}:
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete an uploaded report
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "204":
          description: Report deleted
        "404":
          description: Report not found
        "403":
          description: Editor or administrator access required
        "401":
          description: Authentication required
  /clips:
    get:
      operationId: listClips
      tags: [clips]
      summary: List all clips from the shared workspace archive
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Parsed clips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clip"
        "401":
          description: Authentication required
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get workspace counts and date range
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Workspace summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Authentication required
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: reair_session
  parameters:
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: number
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        createdAt:
          type: string
        isAdmin:
          type: boolean
        role:
          $ref: "#/components/schemas/UserRole"
      required: [id, email, createdAt, isAdmin, role]
    UserRole:
      type: string
      enum: [admin, editor, viewer]
    AuthCredentials:
      type: object
      properties:
        email:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
      required: [email, password]
    CreateUserRequest:
      allOf:
        - $ref: "#/components/schemas/AuthCredentials"
        - type: object
          properties:
            role:
              $ref: "#/components/schemas/UserRole"
          required: [role]
    UpdateUserRoleRequest:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
      required: [role]
    AuthSession:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
    SessionState:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          type: ["object", "null"]
          properties:
            id:
              type: number
            email:
              type: string
            createdAt:
              type: string
            isAdmin:
              type: boolean
            role:
              $ref: "#/components/schemas/UserRole"
          required: [id, email, createdAt, isAdmin, role]
      required: [authenticated, user]
    ReportUpload:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
      required: [name, content]
    Report:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        clipCount:
          type: number
        uploadedAt:
          type: string
      required: [id, name, clipCount, uploadedAt]
    Note:
      type: object
      properties:
        tc:
          type: string
        secs:
          type: ["number", "null"]
        text:
          type: string
      required: [tc, secs, text]
    Clip:
      type: object
      properties:
        id:
          type: string
        reportId:
          type: number
        source:
          type: string
        date:
          type: ["string", "null"]
        revision:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
        originalAir:
          type: ["string", "null"]
        lastAir:
          type: ["string", "null"]
        hosts:
          type: array
          items: { type: string }
        guests:
          type: array
          items: { type: string }
        shortSynopsis:
          type: string
        longSynopsis:
          type: string
        duplicateLongSynopsis:
          type: boolean
        sensitiveNotes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        dateNotes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        flagCount:
          type: number
      required:
        - id
        - reportId
        - source
        - date
        - revision
        - time
        - originalAir
        - lastAir
        - hosts
        - guests
        - shortSynopsis
        - longSynopsis
        - duplicateLongSynopsis
        - sensitiveNotes
        - dateNotes
        - flagCount
    DashboardSummary:
      type: object
      properties:
        reportCount:
          type: number
        clipCount:
          type: number
        reviewCount:
          type: number
        flagCount:
          type: number
        earliestClipDate:
          type: ["string", "null"]
        latestClipDate:
          type: ["string", "null"]
      required:
        - reportCount
        - clipCount
        - reviewCount
        - flagCount
        - earliestClipDate
        - latestClipDate