# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthCredentials } from './authCredentials';
import type { UserRole } from './userRole';

export type CreateUserRequest = AuthCredentials & {
  role: UserRole;
};